Extract class for resolving schema. Fix return values.
